// e2e/frontend/src/api/models.ts
// This file is auto-generated by openapi-zod-ts - do not edit

export interface Order {
  id?: bigint;
  petId?: bigint;
  quantity?: number;
  shipDate?: string; /* date-time */
  status?: "placed" | "approved" | "delivered";
  complete?: boolean;
}

export interface Category {
  id?: bigint;
  name?: string;
}

export interface User {
  id?: bigint;
  username?: string;
  firstName?: string;
  lastName?: string;
  email?: string;
  password?: string;
  phone?: string;
  userStatus?: number;
}

export interface Tag {
  id?: bigint;
  name?: string;
}

export interface Pet {
  id?: bigint;
  name: string;
  category?: Category;
  photoUrls: string[];
  tags?: Tag[];
  status?: "available" | "pending" | "sold";
  microchip_id?: string;
  created_at?: string; /* date-time */
  weight_kg?: number | null;
  attributes?: Record<string, string>;
  external_id?: string | number;
}

export interface PetWritable {
  id?: bigint;
  name: string;
  category?: Category;
  photoUrls: string[];
  tags?: Tag[];
  status?: "available" | "pending" | "sold";
  microchip_id?: string;
  secret_note?: string;
  weight_kg?: number | null;
  attributes?: Record<string, string>;
  external_id?: string | number;
}

export interface ApiResponse {
  code?: number;
  type?: string;
  message?: string;
}